assignment 7
This screenshot shows a region on chromosome 3 in Nothobranchius whitei where different types of genomic data come together to tell a story. The blue peaks show open chromatin from ATAC-seq, meaning those areas of DNA are accessible and could be acting as enhancers or other regulatory elements. The red peaks come from RNA-seq and reflect gene expression. This basically shows which genes are turned on in this region. The purple loops are from Hi-C data and represent parts of the genome that are physically looping together in 3D space. In this view, there's a loop that connects a region with open chromatin (a non-coding OCR) to a spot with active gene expression. This suggests that the non-coding region might be helping control that gene, likely by acting as a distant enhancer.